combine all tests
